feat: add const bs mode
