+++ 2-con/README.md
@@ -1,0 +1,4 @@
+### Usage [WIP]
+```
+make main THREAD_NUM=5 FILTER_TYPE=mb COMPUTE_MODE=by_column
+```
